the signin and signup routes are functioning
